Change author home route to inventory and add inventory by book
