+++ frontend/src/App.jsx
@@ -26,7 +26,6 @@ function App() {
           <Route path="/superadmin/admins" element={<AdminsList />} />
           {/* <Route path='/admin' element={<AdminPage />} /> */}
           {/* <Route path='/author' element={<AuthorPage />} /> */}
-          <Route path="/author" element={<AuthorPage />} />
           <Route path='/admin/authors' element={<AuthorsList />} />
           <Route path="/admin/categories" element={<CategoriesList />} />
           <Route path="/admin/inventories" element={<InventoriesList />} />
