+++ frontend/src/AuthorInventory.css
