+++ frontend/src/ShowInventories.css
